refactor: move WhitelistBlockchainsCard to a folder
